+++ NEWS.md
@@ -1,4 +1,17 @@
-# officedown 0.3.2
+# officedown 0.4.0
+
+## new feature
+
+- plots now accept chunk options `out.width` and `out.height`
+if expressed in percent. #144
+
+# officedown 0.3.3
+
+## issues
+
+- fix manual links for CRAN check
+
+# officedown 0.3.3
 
 ## issues